sraDownloadTruncateCmd: fall back on fastq-dump if wget fails

+++ distiller.nf
@@ -165,6 +165,7 @@ def sraDownloadTruncateCmd(sra_query, library, run, truncate_fastq_reads=0,
         srr_end = (sra_query =~ /end=(\d+)/)[0][1]
     }
 
+    wget_url = "ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR${srrnum.take(3)}/${srr}/${srr}.sra"
     if ((srr_start > 0) || (srr_end != -1)) {
         cmd = """
             ${fastqDumpCmd(srr, library, run, srr_start, srr_end, threads)}
@@ -173,9 +174,15 @@ def sraDownloadTruncateCmd(sra_query, library, run, truncate_fastq_reads=0,
     }
     else {
         cmd = """
-            wget ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR${srrnum.take(3)}/${srr}/${srr}.sra -qO ${srr}.sra
-            ${fastqDumpCmd(srr+'.sra', library, run, 0, -1, threads)}
-            rm ${srr}.sra
+            if wget --spider ${wget_url} 2>/dev/null; then
+                wget ${wget_url} -qO ${srr}.sra
+                ${fastqDumpCmd(srr+'.sra', library, run, 0, -1, threads)}
+                rm ${srr}.sra
+            else
+                echo 'Cannot wget an sra, fall back to fastq-dump'
+                ${fastqDumpCmd(srr, library, run, 0, -1, threads)}
+                if [ -d ./ncbi ]; then rm -Rf ./ncbi; fi
+            fi
         """
     }
